roughly chop context package into separate files. tests not touched, can't quite bear it right now
