+++ worker/uniter/context/context.go
@@ -4,73 +4,27 @@
 package context
 
 import (
-	"bufio"
 	"fmt"
-	"io"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"sort"
-	"strings"
-	"sync"
 	"time"
 
 	"github.com/juju/errors"
 	"github.com/juju/loggo"
 	"github.com/juju/names"
-	utilexec "github.com/juju/utils/exec"
 	"github.com/juju/utils/proxy"
 	"gopkg.in/juju/charm.v4"
 
 	"github.com/juju/juju/api/uniter"
 	"github.com/juju/juju/apiserver/params"
 	"github.com/juju/juju/network"
-	"github.com/juju/juju/version"
-	unitdebug "github.com/juju/juju/worker/uniter/debug"
 	"github.com/juju/juju/worker/uniter/jujuc"
 )
 
 var logger = loggo.GetLogger("juju.worker.uniter.context")
 
-var windowsSuffixOrder = []string{
-	".ps1",
-	".cmd",
-	".bat",
-	".exe",
-}
-
-type missingHookError struct {
-	hookName string
-}
-
-func (e *missingHookError) Error() string {
-	return e.hookName + " does not exist"
-}
-
-func IsMissingHookError(err error) bool {
-	_, ok := err.(*missingHookError)
-	return ok
-}
-
 // meterStatus describes the unit's meter status.
 type meterStatus struct {
 	code string
 	info string
-}
-
-// PortRangeInfo contains information about a pending open- or
-// close-port operation for a port range. This is only exported for
-// testing.
-type PortRangeInfo struct {
-	ShouldOpen  bool
-	RelationTag names.RelationTag
-}
-
-// PortRange contains a port range and a relation id. Used as key to
-// pendingRelations and is only exported for testing.
-type PortRange struct {
-	Ports      network.PortRange
-	RelationId int
 }
 
 // HookContext is the implementation of jujuc.Context.
@@ -355,106 +309,6 @@ func (ctx *HookContext) AddMetrics(key, value string, created time.Time) error {
 	return nil
 }
 
-// mergeEnvironment takes in a string array representing the desired environment
-// and merges it with the current environment. On Windows, clearing the environment,
-// or having missing environment variables, may lead to standard go packages not working
-// (os.TempDir relies on $env:TEMP), and powershell erroring out
-// Currently this function is only used for windows
-func mergeEnvironment(env []string) []string {
-	if env == nil {
-		return nil
-	}
-	m := map[string]string{}
-	var tmpEnv []string
-	for _, val := range os.Environ() {
-		varSplit := strings.SplitN(val, "=", 2)
-		m[varSplit[0]] = varSplit[1]
-	}
-
-	for _, val := range env {
-		varSplit := strings.SplitN(val, "=", 2)
-		m[varSplit[0]] = varSplit[1]
-	}
-
-	for key, val := range m {
-		tmpEnv = append(tmpEnv, key+"="+val)
-	}
-
-	return tmpEnv
-}
-
-// windowsEnv adds windows specific environment variables. PSModulePath
-// helps hooks use normal imports instead of dot sourcing modules
-// its a convenience variable. The PATH variable delimiter is
-// a semicolon instead of a colon
-func (ctx *HookContext) windowsEnv(charmDir, toolsDir string) []string {
-	charmModules := filepath.Join(charmDir, "Modules")
-	hookModules := filepath.Join(charmDir, "hooks", "Modules")
-	env := []string{
-		"Path=" + filepath.FromSlash(toolsDir) + ";" + os.Getenv("Path"),
-		"PSModulePath=" + os.Getenv("PSModulePath") + ";" + charmModules + ";" + hookModules,
-	}
-	return mergeEnvironment(env)
-}
-
-func (ctx *HookContext) ubuntuEnv(toolsDir string) []string {
-	env := []string{
-		"APT_LISTCHANGES_FRONTEND=none",
-		"DEBIAN_FRONTEND=noninteractive",
-		"PATH=" + toolsDir + ":" + os.Getenv("PATH"),
-	}
-	return env
-}
-
-func (ctx *HookContext) osDependentEnvVars(charmDir, toolsDir string) []string {
-	switch version.Current.OS {
-	case version.Windows:
-		return ctx.windowsEnv(charmDir, toolsDir)
-	default:
-		return ctx.ubuntuEnv(toolsDir)
-	}
-}
-
-// hookVars returns an os.Environ-style list of strings necessary to run a hook
-// such that it can know what environment it's operating in, and can call back
-// into ctx.
-func (ctx *HookContext) hookVars(charmDir, toolsDir, socketPath string) []string {
-	// TODO(binary132): add Action env variables: JUJU_ACTION_NAME,
-	// JUJU_ACTION_UUID, ...
-	vars := []string{
-		"CHARM_DIR=" + charmDir,
-		"JUJU_CONTEXT_ID=" + ctx.id,
-		"JUJU_AGENT_SOCKET=" + socketPath,
-		"JUJU_UNIT_NAME=" + ctx.unit.Name(),
-		"JUJU_ENV_UUID=" + ctx.uuid,
-		"JUJU_ENV_NAME=" + ctx.envName,
-		"JUJU_API_ADDRESSES=" + strings.Join(ctx.apiAddrs, " "),
-	}
-	osVars := ctx.osDependentEnvVars(charmDir, toolsDir)
-	vars = append(vars, osVars...)
-
-	if r, found := ctx.HookRelation(); found {
-		vars = append(vars, "JUJU_RELATION="+r.Name())
-		vars = append(vars, "JUJU_RELATION_ID="+r.FakeId())
-		name, _ := ctx.RemoteUnitName()
-		vars = append(vars, "JUJU_REMOTE_UNIT="+name)
-	}
-	vars = append(vars, ctx.proxySettings.AsEnvironmentValues()...)
-	vars = append(vars, ctx.meterStatusEnvVars()...)
-	return vars
-}
-
-// meterStatusEnvVars returns meter status environment variables if the meter
-// status is set.
-func (ctx *HookContext) meterStatusEnvVars() []string {
-	if ctx.meterStatus != nil {
-		return []string{
-			fmt.Sprintf("JUJU_METER_STATUS=%s", ctx.meterStatus.code),
-			fmt.Sprintf("JUJU_METER_INFO=%s", ctx.meterStatus.info)}
-	}
-	return nil
-}
-
 func (ctx *HookContext) finalizeContext(process string, ctxErr error) (err error) {
 	writeChanges := ctxErr == nil
 
@@ -569,513 +423,4 @@ func (ctx *HookContext) finalizeAction(err, unhandledErr error) error {
 		unhandledErr = errors.Wrap(unhandledErr, callErr)
 	}
 	return unhandledErr
-}
-
-// RunCommands executes the commands in an environment which allows it to to
-// call back into the hook context to execute jujuc tools.
-func (ctx *HookContext) RunCommands(commands, charmDir, toolsDir, socketPath string) (*utilexec.ExecResponse, error) {
-	env := ctx.hookVars(charmDir, toolsDir, socketPath)
-	result, err := utilexec.RunCommands(
-		utilexec.RunParams{
-			Commands:    commands,
-			WorkingDir:  charmDir,
-			Environment: env})
-	return result, ctx.finalizeContext("run commands", err)
-}
-
-func (ctx *HookContext) GetLogger(hookName string) loggo.Logger {
-	return loggo.GetLogger(fmt.Sprintf("unit.%s.%s", ctx.UnitName(), hookName))
-}
-
-// RunAction executes a hook from the charm's actions in an environment which
-// allows it to to call back into the hook context to execute jujuc tools.
-func (ctx *HookContext) RunAction(hookName, charmDir, toolsDir, socketPath string) error {
-	if ctx.actionData == nil {
-		return fmt.Errorf("not running an action")
-	}
-	// If the action had already failed (i.e. from invalid params), we
-	// just want to finalize without running it.
-	if ctx.actionData.ActionFailed {
-		return ctx.finalizeContext(hookName, nil)
-	}
-	return ctx.runCharmHookWithLocation(hookName, "actions", charmDir, toolsDir, socketPath)
-}
-
-// RunHook executes a built-in hook in an environment which allows it to to
-// call back into the hook context to execute jujuc tools.
-func (ctx *HookContext) RunHook(hookName, charmDir, toolsDir, socketPath string) error {
-	return ctx.runCharmHookWithLocation(hookName, "hooks", charmDir, toolsDir, socketPath)
-}
-
-func (ctx *HookContext) runCharmHookWithLocation(hookName, charmLocation, charmDir, toolsDir, socketPath string) error {
-	var err error
-	env := ctx.hookVars(charmDir, toolsDir, socketPath)
-	debugctx := unitdebug.NewHooksContext(ctx.unit.Name())
-	if session, _ := debugctx.FindSession(); session != nil && session.MatchHook(hookName) {
-		logger.Infof("executing %s via debug-hooks", hookName)
-		err = session.RunHook(hookName, charmDir, env)
-	} else {
-		err = ctx.runCharmHook(hookName, charmDir, env, charmLocation)
-	}
-	return ctx.finalizeContext(hookName, err)
-}
-
-func lookPath(hook string) (string, error) {
-	hookFile, err := exec.LookPath(hook)
-	if err != nil {
-		if ee, ok := err.(*exec.Error); ok && os.IsNotExist(ee.Err) {
-			return "", &missingHookError{hook}
-		}
-		return "", err
-	}
-	return hookFile, nil
-}
-
-// searchHook will search, in order, hooks suffixed with extensions
-// in windowsSuffixOrder. As windows cares about extensions to determine
-// how to execute a file, we will allow several suffixes, with powershell
-// being default.
-func searchHook(charmDir, hook string) (string, error) {
-	hookFile := filepath.Join(charmDir, hook)
-	if version.Current.OS != version.Windows {
-		// we are not running on windows,
-		// there is no need to look for suffixed hooks
-		return lookPath(hookFile)
-	}
-	for _, val := range windowsSuffixOrder {
-		file := fmt.Sprintf("%s%s", hookFile, val)
-		foundHook, err := lookPath(file)
-		if err != nil {
-			if IsMissingHookError(err) {
-				// look for next suffix
-				continue
-			}
-			return "", err
-		}
-		return foundHook, nil
-	}
-	return "", &missingHookError{hook}
-}
-
-// hookCommand constructs an appropriate command to be passed to
-// exec.Command(). The exec package uses cmd.exe as default on windows
-// cmd.exe does not know how to execute ps1 files by default, and
-// powershell needs a few flags to allow execution (-ExecutionPolicy)
-// and propagate error levels (-File). .cmd and .bat files can be run directly
-func hookCommand(hook string) []string {
-	if version.Current.OS != version.Windows {
-		// we are not running on windows,
-		// just return the hook name
-		return []string{hook}
-	}
-	if strings.HasSuffix(hook, ".ps1") {
-		return []string{
-			"powershell.exe",
-			"-NonInteractive",
-			"-ExecutionPolicy",
-			"RemoteSigned",
-			"-File",
-			hook,
-		}
-	}
-	return []string{hook}
-}
-
-func (ctx *HookContext) runCharmHook(hookName, charmDir string, env []string, charmLocation string) error {
-	hook, err := searchHook(charmDir, filepath.Join(charmLocation, hookName))
-	if err != nil {
-		if IsMissingHookError(err) {
-			// Missing hook is perfectly valid, but worth mentioning.
-			logger.Infof("skipped %q hook (not implemented)", hookName)
-		}
-		return err
-	}
-	hookCmd := hookCommand(hook)
-	ps := exec.Command(hookCmd[0], hookCmd[1:]...)
-	ps.Env = env
-	ps.Dir = charmDir
-	outReader, outWriter, err := os.Pipe()
-	if err != nil {
-		return fmt.Errorf("cannot make logging pipe: %v", err)
-	}
-	ps.Stdout = outWriter
-	ps.Stderr = outWriter
-	hookLogger := &hookLogger{
-		r:      outReader,
-		done:   make(chan struct{}),
-		logger: ctx.GetLogger(hookName),
-	}
-	go hookLogger.run()
-	err = ps.Start()
-	outWriter.Close()
-	if err == nil {
-		err = ps.Wait()
-	}
-	hookLogger.stop()
-	return err
-}
-
-type hookLogger struct {
-	r       io.ReadCloser
-	done    chan struct{}
-	mu      sync.Mutex
-	stopped bool
-	logger  loggo.Logger
-}
-
-func (l *hookLogger) run() {
-	defer close(l.done)
-	defer l.r.Close()
-	br := bufio.NewReaderSize(l.r, 4096)
-	for {
-		line, _, err := br.ReadLine()
-		if err != nil {
-			if err != io.EOF {
-				logger.Errorf("cannot read hook output: %v", err)
-			}
-			break
-		}
-		l.mu.Lock()
-		if l.stopped {
-			l.mu.Unlock()
-			return
-		}
-		l.logger.Infof("%s", line)
-		l.mu.Unlock()
-	}
-}
-
-func (l *hookLogger) stop() {
-	// We can see the process exit before the logger has processed
-	// all its output, so allow a moment for the data buffered
-	// in the pipe to be processed. We don't wait indefinitely though,
-	// because the hook may have started a background process
-	// that keeps the pipe open.
-	select {
-	case <-l.done:
-	case <-time.After(100 * time.Millisecond):
-	}
-	// We can't close the pipe asynchronously, so just
-	// stifle output instead.
-	l.mu.Lock()
-	l.stopped = true
-	l.mu.Unlock()
-}
-
-// SettingsMap is a map from unit name to relation settings.
-type SettingsMap map[string]params.RelationSettings
-
-// ContextRelation is the implementation of jujuc.ContextRelation.
-type ContextRelation struct {
-	ru *uniter.RelationUnit
-
-	// members contains settings for known relation members. Nil values
-	// indicate members whose settings have not yet been cached.
-	members SettingsMap
-
-	// settings allows read and write access to the relation unit settings.
-	settings *uniter.Settings
-
-	// cache is a short-term cache that enables consistent access to settings
-	// for units that are not currently participating in the relation. Its
-	// contents should be cleared whenever a new hook is executed.
-	cache SettingsMap
-}
-
-// NewContextRelation creates a new context for the given relation unit.
-// The unit-name keys of members supplies the initial membership.
-func NewContextRelation(ru *uniter.RelationUnit, members map[string]int64) *ContextRelation {
-	ctx := &ContextRelation{ru: ru, members: SettingsMap{}}
-	for unit := range members {
-		ctx.members[unit] = nil
-	}
-	ctx.ClearCache()
-	return ctx
-}
-
-// WriteSettings persists all changes made to the unit's relation settings.
-func (ctx *ContextRelation) WriteSettings() (err error) {
-	if ctx.settings != nil {
-		err = ctx.settings.Write()
-	}
-	return
-}
-
-// ClearCache discards all cached settings for units that are not members
-// of the relation, and all unwritten changes to the unit's relation settings.
-// including any changes to Settings that have not been written.
-func (ctx *ContextRelation) ClearCache() {
-	ctx.settings = nil
-	ctx.cache = make(SettingsMap)
-}
-
-// UpdateMembers ensures that the context is aware of every supplied
-// member unit. For each supplied member, the cached settings will be
-// overwritten.
-func (ctx *ContextRelation) UpdateMembers(members SettingsMap) {
-	for m, s := range members {
-		ctx.members[m] = s
-	}
-}
-
-// DeleteMember drops the membership and cache of a single remote unit, without
-// perturbing settings for the remaining members.
-func (ctx *ContextRelation) DeleteMember(unitName string) {
-	delete(ctx.members, unitName)
-}
-
-func (ctx *ContextRelation) Id() int {
-	return ctx.ru.Relation().Id()
-}
-
-func (ctx *ContextRelation) Name() string {
-	return ctx.ru.Endpoint().Name
-}
-
-func (ctx *ContextRelation) FakeId() string {
-	return fmt.Sprintf("%s:%d", ctx.Name(), ctx.ru.Relation().Id())
-}
-
-func (ctx *ContextRelation) UnitNames() (units []string) {
-	for unit := range ctx.members {
-		units = append(units, unit)
-	}
-	sort.Strings(units)
-	return units
-}
-
-func (ctx *ContextRelation) Settings() (jujuc.Settings, error) {
-	if ctx.settings == nil {
-		node, err := ctx.ru.Settings()
-		if err != nil {
-			return nil, err
-		}
-		ctx.settings = node
-	}
-	return ctx.settings, nil
-}
-
-func (ctx *ContextRelation) ReadSettings(unit string) (settings params.RelationSettings, err error) {
-	settings, member := ctx.members[unit]
-	if settings == nil {
-		if settings = ctx.cache[unit]; settings == nil {
-			settings, err = ctx.ru.ReadSettings(unit)
-			if err != nil {
-				return nil, err
-			}
-		}
-	}
-	if member {
-		ctx.members[unit] = settings
-	} else {
-		ctx.cache[unit] = settings
-	}
-	return settings, nil
-}
-
-// ActionData contains the tag, parameters, and results of an Action.
-type ActionData struct {
-	ActionTag      names.ActionTag
-	ActionParams   map[string]interface{}
-	ActionFailed   bool
-	ResultsMessage string
-	ResultsMap     map[string]interface{}
-}
-
-// newActionData builds a suitable actionData struct with no nil members.
-// this should only be called in the event that an Action hook is being requested.
-func NewActionData(tag *names.ActionTag, params map[string]interface{}) *ActionData {
-	return &ActionData{
-		ActionTag:    *tag,
-		ActionParams: params,
-		ResultsMap:   map[string]interface{}{},
-	}
-}
-
-// actionStatus messages define the possible states of a completed Action.
-const (
-	actionStatusInit   = "init"
-	actionStatusFailed = "fail"
-)
-
-// addValueToMap adds the given value to the map on which the method is run.
-// This allows us to merge maps such as {foo: {bar: baz}} and {foo: {baz: faz}}
-// into {foo: {bar: baz, baz: faz}}.
-func addValueToMap(keys []string, value string, target map[string]interface{}) {
-	next := target
-
-	for i := range keys {
-		// if we are on last key set the value.
-		// shouldn't be a problem.  overwrites existing vals.
-		if i == len(keys)-1 {
-			next[keys[i]] = value
-			break
-		}
-
-		if iface, ok := next[keys[i]]; ok {
-			switch typed := iface.(type) {
-			case map[string]interface{}:
-				// If we already had a map inside, keep
-				// stepping through.
-				next = typed
-			default:
-				// If we didn't, then overwrite value
-				// with a map and iterate with that.
-				m := map[string]interface{}{}
-				next[keys[i]] = m
-				next = m
-			}
-			continue
-		}
-
-		// Otherwise, it wasn't present, so make it and step
-		// into.
-		m := map[string]interface{}{}
-		next[keys[i]] = m
-		next = m
-	}
-}
-
-func validatePortRange(protocol string, fromPort, toPort int) (network.PortRange, error) {
-	// Validate the given range.
-	newRange := network.PortRange{
-		Protocol: strings.ToLower(protocol),
-		FromPort: fromPort,
-		ToPort:   toPort,
-	}
-	if err := newRange.Validate(); err != nil {
-		return network.PortRange{}, err
-	}
-	return newRange, nil
-}
-
-func tryOpenPorts(
-	protocol string,
-	fromPort, toPort int,
-	unitTag names.UnitTag,
-	machinePorts map[network.PortRange]params.RelationUnit,
-	pendingPorts map[PortRange]PortRangeInfo,
-) error {
-	// TODO(dimitern) Once port ranges are linked to relations in
-	// addition to networks, refactor this functions and test it
-	// better to ensure it handles relations properly.
-	relationId := -1
-
-	//Validate the given range.
-	newRange, err := validatePortRange(protocol, fromPort, toPort)
-	if err != nil {
-		return err
-	}
-	rangeKey := PortRange{
-		Ports:      newRange,
-		RelationId: relationId,
-	}
-
-	rangeInfo, isKnown := pendingPorts[rangeKey]
-	if isKnown {
-		if !rangeInfo.ShouldOpen {
-			// If the same range is already pending to be closed, just
-			// mark is pending to be opened.
-			rangeInfo.ShouldOpen = true
-			pendingPorts[rangeKey] = rangeInfo
-		}
-		return nil
-	}
-
-	// Ensure there are no conflicts with existing ports on the
-	// machine.
-	for portRange, relUnit := range machinePorts {
-		relUnitTag, err := names.ParseUnitTag(relUnit.Unit)
-		if err != nil {
-			return errors.Annotatef(
-				err,
-				"machine ports %v contain invalid unit tag",
-				portRange,
-			)
-		}
-		if newRange.ConflictsWith(portRange) {
-			if portRange == newRange && relUnitTag == unitTag {
-				// The same unit trying to open the same range is just
-				// ignored.
-				return nil
-			}
-			return errors.Errorf(
-				"cannot open %v (unit %q): conflicts with existing %v (unit %q)",
-				newRange, unitTag.Id(), portRange, relUnitTag.Id(),
-			)
-		}
-	}
-	// Ensure other pending port ranges do not conflict with this one.
-	for rangeKey, rangeInfo := range pendingPorts {
-		if newRange.ConflictsWith(rangeKey.Ports) && rangeInfo.ShouldOpen {
-			return errors.Errorf(
-				"cannot open %v (unit %q): conflicts with %v requested earlier",
-				newRange, unitTag.Id(), rangeKey.Ports,
-			)
-		}
-	}
-
-	rangeInfo = pendingPorts[rangeKey]
-	rangeInfo.ShouldOpen = true
-	pendingPorts[rangeKey] = rangeInfo
-	return nil
-}
-
-func tryClosePorts(
-	protocol string,
-	fromPort, toPort int,
-	unitTag names.UnitTag,
-	machinePorts map[network.PortRange]params.RelationUnit,
-	pendingPorts map[PortRange]PortRangeInfo,
-) error {
-	// TODO(dimitern) Once port ranges are linked to relations in
-	// addition to networks, refactor this functions and test it
-	// better to ensure it handles relations properly.
-	relationId := -1
-
-	// Validate the given range.
-	newRange, err := validatePortRange(protocol, fromPort, toPort)
-	if err != nil {
-		return err
-	}
-	rangeKey := PortRange{
-		Ports:      newRange,
-		RelationId: relationId,
-	}
-
-	rangeInfo, isKnown := pendingPorts[rangeKey]
-	if isKnown {
-		if rangeInfo.ShouldOpen {
-			// If the same range is already pending to be opened, just
-			// remove it from pending.
-			delete(pendingPorts, rangeKey)
-		}
-		return nil
-	}
-
-	// Ensure the range we're trying to close is opened on the
-	// machine.
-	relUnit, found := machinePorts[newRange]
-	if !found {
-		// Trying to close a range which is not open is ignored.
-		return nil
-	} else if relUnit.Unit != unitTag.String() {
-		relUnitTag, err := names.ParseUnitTag(relUnit.Unit)
-		if err != nil {
-			return errors.Annotatef(
-				err,
-				"machine ports %v contain invalid unit tag",
-				newRange,
-			)
-		}
-		return errors.Errorf(
-			"cannot close %v (opened by %q) from %q",
-			newRange, relUnitTag.Id(), unitTag.Id(),
-		)
-	}
-
-	rangeInfo = pendingPorts[rangeKey]
-	rangeInfo.ShouldOpen = false
-	pendingPorts[rangeKey] = rangeInfo
-	return nil
 }
